# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SIAA — Sistema Integrado de Asistencia Académica
  version: 0.1.0
  description: |
    API REST de SIAA. Todos los endpoints requieren HTTPS. Las respuestas de
    error siguen la estructura del catálogo §9.2 del SRS.
  contact:
    name: Equipo SIAA
    email: dev@siaa.edu.co
  license:
    name: Propietario
servers:
  - url: https://api.siaa.edu.co/api/v1
    description: Producción
  - url: http://localhost:8080/api/v1
    description: Desarrollo local

# ─── Seguridad global ──────────────────────────────────────────────────────
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Esquemas de error comunes ─────────────────────────────────────────
  schemas:
    ErrorResponse:
      type: object
      required: [codigo, mensaje, correlationId]
      properties:
        codigo:
          type: string
          example: AUTH_CREDENCIALES_INVALIDAS
          description: Código de error del catálogo §9.2
        mensaje:
          type: string
          example: Credenciales incorrectas
        detalles:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        correlationId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
    FieldError:
      type: object
      required: [campo, error]
      properties:
        campo:
          type: string
          example: correo
        error:
          type: string
          example: required

    # ─── Auth ──────────────────────────────────────────────────────────
...
    TokenPairResponse:
      type: object
      required: [accessToken, refreshToken, expiraEn, tipoToken, usuario]
      properties:
        accessToken:
          type: string
          description: JWT de acceso válido por 15 minutos
          example: eyJhbGciOiJIUzI1NiJ9...
        refreshToken:
          type: string
          description: Token de refresco opaco válido por 30 días
          example: a1b2c3d4e5f6...
        expiraEn:
          type: string
          format: date-time
          example: "2024-12-31T23:59:59Z"
        tipoToken:
          type: string
          enum: [Bearer]
        usuario:
          $ref: '#/components/schemas/UsuarioInfo'
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
        dispositivoId:
          type: string
...
    UsuarioInfo:
      type: object
      required: [id, correo, nombre, apellido, roles, permisos]
      properties:
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        correo:
          type: string
          format: email
        nombre:
          type: string
        apellido:
          type: string
        roles:
          type: array
          items:
            type: string
          example: [DOCENTE]
        permisos:
          type: array
          items:
            type: string
          example: [marcaje:crear, marcaje:leer]
...
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Renovar par de tokens
      description: |
        Rota el refresh token. Si se detecta reutilización, se revoca toda la
        familia de tokens (AC-06).
      security: []
      tags: [autenticación]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Nuevo par de tokens emitido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairResponse'
        '401':
          description: Token inválido, expirado o revocado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /sesiones/{id}:
    get:
      operationId: obtenerSesionPorId
      summary: Obtener detalle de una sesión de clase (US-ACA-05)
      tags: [sesiones]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalle de sesión
        '404':
          description: Sesión no encontrada
  /sesiones/{id}/cancelar:
    post:
      operationId: cancelarSesion
      summary: Cancelar administrativamente una sesión de clase (US-ACA-08)
      tags: [sesiones]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [motivo]
              properties:
                motivo:
                  type: string
      responses:
        '200':
          description: Sesión cancelada exitosamente
        '401':
          description: No autenticado
        '404':
          description: Sesión no encontrada
  /sesiones/{id}/aula:
    put:
      operationId: reasignarAulaSesion
      summary: Reasignar aula para una sesión puntual congelando su geometría (US-ACA-06)
      tags: [sesiones]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [espacioId, motivo]
              properties:
                espacioId:
                  type: string
                motivo:
                  type: string
      responses:
        '200':
          description: Aula reasignada exitosamente
        '401':
          description: No autenticado
        '404':
          description: Sesión o aula no encontrada
  /academico/importar/preview:
    post:
      operationId: previewImportarAcademicoCSV
      summary: Validar y previsualizar archivo CSV de estructura académica (US-ACA-07)
      tags: [academico]
      responses:
        '200':
          description: Preview con errores por fila y resumen
        '400':
          description: Archivo CSV inválido o vacío
  /academico/importar:
    post:
      operationId: confirmarImportarAcademico
      summary: Confirmar importación masiva de asignaciones académicas (US-ACA-07)
      tags: [academico]
      responses:
        '201':
          description: Asignaciones creadas e importadas
        '400':
          description: Filas vacías o inválidas
...